Improve openstack-release file discovery
In cases where the deployment node is seperated from the infrastructure
node, the openstack-release file discovery logic might calculate the
wrong release since the /etc/openstack-release file does not get generated
on the deployment node.
This commit adds logic to pull down the /etc/openstack-release file from
the infrastucture node to the deployment node. This way, the release discovery
logic will properly assess the openstack release even in situations where the
deployment node is isolated from the infra node.
If the openstack-release file is not found, an attempt to retrieve the
rpc-release file is made.

A conditional is added to check if the rpc-release file exist in the
environment. If it does, and there is no openstack-release file, then
the deployment version is set to Juno.

Change-Id: I6136609ef11be7b98ddcfc7589dd17cf1e98c362